treewide: clean up the tree
so you might be wondering why all of this garbage was here to begin with. i asked myself this too, and the answer is i don't know. past fakeshell was probably working on something and added a bunch of stuff as potential fixes and never ended up cleaning any of it up. but now that we're here lets clean it up anyways

Signed-off-by: Bardia Moshiri <fakeshell@bardia.tech>